backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pymysql
cryptography
anthropic
boto3
python-multipart
PyMuPDF
python-docx
openpyxl
pillow
pandas
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
msal
alembic
